renaming regina test to regina auto test as we'll be creating regina unit tests soon and we don't want ambiguity
