+++ config/tests/regina_auto_test/code_generation.json
