+++ tests/regina_auto_test/src/regina_boot_test.cpp
@@ -1,9 +1,9 @@
-#include "regina_test/regina_boot_test.h"
+#include "regina_auto_test/regina_boot_test.h"
 #include "regina/regina.h"
 #include "rex_engine/event_system.h"
 #include "rex_windows/platform_creation_params.h"
 
-namespace regina_test
+namespace regina_auto_test
 {
   bool initialize()
   {
@@ -26,10 +26,10 @@ namespace regina_test
 
     app_params.gui_params.window_title = "Regina Test";
 
-    app_params.engine_params.app_init_func = regina_test::initialize;
-    app_params.engine_params.app_update_func = regina_test::update;
-    app_params.engine_params.app_shutdown_func = regina_test::shutdown;
+    app_params.engine_params.app_init_func = regina_auto_test::initialize;
+    app_params.engine_params.app_update_func = regina_auto_test::update;
+    app_params.engine_params.app_shutdown_func = regina_auto_test::shutdown;
 
     return app_params;
   }
-} // namespace regina_test
+} // namespace regina_auto_test
